Build a new NCE module
NCE should not only serve as a simple `Loss` layer in pytorch, because
NCE requires to cooperate with the decoder(a linear layer). The decoder
is migrated from RNN module to NCE module in hope to reduce unnecessary
decoding computation.

The actual speed of this version is pretty slow, unacceptable
100s/batch. Need more optimization.