Refactor the NCE loss framework and index module
The index module was a sub-module before. The sub-module doesn't know
the property of its' main module, where the important properties such as
normalization term and noise ratio are stored.

In this commit, I made the NCELoss a base class which handles most dirty
work like noise generating and NCE loss computing.